+++ assets/src/apps/authoring/components/AdaptivityEditor/ConditionsBlockEditor.tsx
@@ -258,7 +258,7 @@ const ConditionsBlockEditor: React.FC<CondtionsBlockEditorProps> = (props) => {
           data-toggle="dropdown"
           aria-haspopup="true"
           aria-expanded="false"
-          onClick={(e) => {
+          onClick={() => {
             ($(`#cb-editor-add-context-trigger`) as any).dropdown('toggle');
           }}
         >
@@ -333,8 +333,7 @@ const ConditionsBlockEditor: React.FC<CondtionsBlockEditorProps> = (props) => {
                   data-toggle="dropdown"
                   aria-haspopup="true"
                   aria-expanded="false"
-                  onClick={(e) => {
-                    // e.stopPropagation();
+                  onClick={() => {
                     ($(`#add-condition-${id}-context-trigger-${index}`) as any).dropdown('toggle');
                   }}
                 >