+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 ### Bug Fixes
 
 - Fix a style issue with the workspace footer
+- Prevent objectives used in selections from being deleted
 - Fix an issue where modals misbehaved sporadically
+- Move "Many Students Wonder" from activity styling to content styling
+- Fix an issue where nonstructural section resources were missing after update
 
 ### Enhancements
 
@@ -14,6 +17,11 @@
 - Add advanced section creation remix
 - Allow for section creation from course products
 - Add analytics / insights data export button
+- Add ability for an admin to browse all course sections
+- Add server driven paged, sortable table for project list
+- Add ability to remix materials from multiple projects
+- Fix insert content popup in page editor
+
 
 ### Release Notes
 
@@ -26,12 +34,15 @@ STRIPE_PUBLIC_SECRET    (Required if PAYMENT_PROVIDER=stripe)
 STRIPE_PRIVATE_SECRET   (Required if PAYMENT_PROVIDER=stripe)
 ```
 
+<<<<<<< HEAD
 ## 0.13.8 (2021-10-07)
 
 ### Bug Fixes
 
 - Handle titles of activities correctly in analytics download
 
+=======
+>>>>>>> master
 ## 0.13.7 (2021-10-06)
 
 ### Bug Fixes